UPDATE: 2nd testing bug fix

+++ src/hooks/useMyMemo.ts
@@ -1,4 +1,4 @@
-import { useCallback } from "react";
+import { useState, useCallback } from "react";
 
 import type {
   PostMemosRequest,
@@ -8,7 +8,11 @@ import type {
 // const userId = "55a0ef11-c9c8-471d-adeb-29b87d3d6bdc";
 
 export default function useMyMemo() {
+  const [loading, setLoading] = useState(false);
+
   const postMemos = useCallback(async (data: PostMemosRequest) => {
+    if (loading) return;
+    setLoading(true);
     const jsonData = JSON.stringify(data);
     const res = await fetch(`/api/memos`, {
       method: "POST",
@@ -18,10 +22,13 @@ export default function useMyMemo() {
       const body = await res.json();
       throw new Error(body.error);
     }
+    setLoading(false);
     return;
   }, []);
 
   const updateMemos = useCallback(async (data: UpdateMemosRequest) => {
+    if (loading) return;
+    setLoading(true);
     const jsonData = JSON.stringify(data);
     const res = await fetch(`/api/memos`, {
       method: "PUT",
@@ -31,10 +38,13 @@ export default function useMyMemo() {
       const body = await res.json();
       throw new Error(body.error);
     }
+    setLoading(false);
     return;
   }, []);
 
   const deleteMemos = useCallback(async (deletedIdArray: string[]) => {
+    if (loading) return;
+    setLoading(true);
     const deletedIdArrayToString = deletedIdArray.join();
     const res = await fetch(`/api/memos`, {
       method: "DELETE",
@@ -44,6 +54,7 @@ export default function useMyMemo() {
       const body = await res.json();
       throw new Error(body.error);
     }
+    setLoading(false);
     return;
   }, []);
